back - move warnIfUserCreatesTooManyCompanies into createCompany mutation
