+++ back/src/companies/mutations/__tests__/too-many-companies.test.ts
@@ -1,0 +1,44 @@
+import { User, Company } from "../../../generated/prisma-client";
+import { warnIfUserCreatesTooManyCompanies } from "../create-company";
+
+const countMock = jest.fn();
+const mailMock = jest.fn();
+
+jest.mock("../../../generated/prisma-client", () => ({
+  prisma: {
+    companyAssociationsConnection: () => ({
+      aggregate: () => ({ count: countMock })
+    })
+  }
+}));
+
+jest.mock("../../../common/mails.helper", () => ({
+  sendMail: () => mailMock()
+}));
+
+describe("warnIfUserCreatesTooManyCompanies subscription", () => {
+  beforeEach(() => {
+    jest.resetAllMocks();
+  });
+
+  test("should send mail if user has too much companies", async () => {
+    // 10 > MAX
+    countMock.mockResolvedValue(10);
+    await warnIfUserCreatesTooManyCompanies(
+      { id: "id", name: "name" } as User,
+      { name: "companyName", siret: "siret" } as Company
+    );
+    expect(mailMock).toHaveBeenCalledTimes(1);
+  });
+
+  test("should not send mail if user has low number of companies", async () => {
+    // 1 < MAX
+    countMock.mockResolvedValue(1);
+    await warnIfUserCreatesTooManyCompanies(
+      { id: "id", name: "name" } as User,
+      { name: "companyName", siret: "siret" } as Company
+    );
+
+    expect(mailMock).not.toHaveBeenCalled();
+  });
+});
